adding twss command

+++ src/discordUtils.ts
@@ -4,9 +4,10 @@ import Utils from './utils';
 export default class DiscordUtils {
     static serverId = '614956907261722687';
 
-    static async sendChannelMessage (msg: Message, reply: string): Promise<void> {
+    static async sendChannelMessage (msg: Message, reply: string, del = true): Promise<void> {
         msg.channel.send(reply);
-        await msg.delete();
+        if (del)
+            await msg.delete();
     }
 
     static async sendReply (msg: Message, reply: string): Promise<void> {
